reorganize pyvelox code
