--- MANIFEST.in
@@ -1,8 +1,0 @@
-include *.txt
-include *.cmake
-graft CMake
-graft third_party
-graft velox
-graft pyvelox
-graft scripts
-prune velox/docs
