+++ python/legacy/signatures.cpp
