+++ python/legacy/signatures.h
