+++ python/plan_builder/PyPlanBuilder.h
@@ -18,8 +18,8 @@
 
 #include "velox/core/PlanNode.h"
 #include "velox/exec/tests/utils/PlanBuilder.h"
-#include "velox/py/file/PyFile.h"
-#include "velox/py/type/PyType.h"
+#include "python/file/PyFile.h"
+#include "python/type/PyType.h"
 
 namespace facebook::velox::py {
