+++ python/test/test_vector.py
