+++ python/vector/PyVector.cpp
@@ -14,11 +14,26 @@
  * limitations under the License.
  */
 
-#include "velox/py/type/PyType.h"
-#include <pybind11/stl.h>
+#include "python/vector/PyVector.h"
+#include "velox/vector/ComplexVector.h"
+#include "velox/vector/VectorPrinter.h"
 
 namespace facebook::velox::py {
 
-namespace py = pybind11;
+std::string PyVector::summarizeToText() const {
+  return velox::VectorPrinter::summarizeToText(*vector_);
+}
+
+std::string PyVector::printDetailed() const {
+  return velox::printVector(*vector_);
+}
+
+PyVector PyVector::childAt(vector_size_t idx) const {
+  if (auto rowVector = std::dynamic_pointer_cast<RowVector>(vector_)) {
+    return PyVector{rowVector->childAt(idx), pool_};
+  }
+  throw std::runtime_error(fmt::format(
+      "Can only call child_at() on RowVector, but got '{}'", toString()));
+}
 
 } // namespace facebook::velox::py
