+++ velox/CMakeLists.txt
@@ -75,10 +75,6 @@ if(${VELOX_ENABLE_GPU})
   add_subdirectory(external/jitify)
 endif()
 
-if(${VELOX_BUILD_PYTHON_PACKAGE} AND NOT VELOX_PYTHON_LEGACY_ONLY)
-  add_subdirectory(py)
-endif()
-
 # substrait converter
 if(${VELOX_ENABLE_SUBSTRAIT})
   add_subdirectory(substrait)
